Iteration 0: make a demo that works
It is only capable of connecting to non-jwt connections
over HTTP, but at least it prints results.
The current iteration has concurrency of 1 and is based
on a very naive solution that keeps the Hrana Client
under an Arc+Mutex. Next iteration will properly implement
a reactor capable of sending requests concurrently and matching
them with appropriate responses with the same id.