Iteration 1: move to reactor scheme
Iteration 1 drops the Arc<Mutex<>> and instead moves
the I/O reactor to a separate fiber. This fiber communicates
with Hrana streams via channels. It's still just as bad,
because it only handles one sequence of send->receive,
instead of having a separate reader and writer fibers,
but it's one step closer to the final design, where
there are two fibers:
1. One for sending requests to the server
2. One for receiving requests from the server
   and routing them to appropriate streams