Iteration 2: implement separate reader and writer fibers
The fibers are responsible for sending messages to the server
and receiving them. The fibers communicate with stream objects
by channels.

One important TODO left is to remove the busy loop from
the reader fiber. Instead, it should block until any of the channels
has available data, and then communicate with the server.