Iteration 3: switch to tokio::sync::Notify
From now on, after iterating through the per-stream channels
and trying to check if there are any pending requests to forward
to the server, the write fiber is going to go to sleep on a Notify
instance. Any call of Stream::execute() will send a notification
and wake up the writer fiber.